в форму отображения тегов добавлена строка с расположением песни, добавлено отображение размера

+++ readertag.h
@@ -25,7 +25,7 @@ typedef struct TagInfo{
     int length;
     int beatRate;
     int year;    
-    int size;
+    quint64 size;
     int frequency;
 } TagInfo;
